+++ packages/orq-rc/RELEASES.md
@@ -11879,3 +11879,13 @@ Based on:
 - [typescript v4.14.0-rc.52] packages/orq-rc
 ### Releases
 - [NPM v4.14.0-rc.52] https://www.npmjs.com/package/@orq-ai/node/v/4.14.0-rc.52 - packages/orq-rc
+
+## 2026-08-25 12:37:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.795.3 (2.932.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.14.0-rc.53] packages/orq-rc
+### Releases
+- [NPM v4.14.0-rc.53] https://www.npmjs.com/package/@orq-ai/node/v/4.14.0-rc.53 - packages/orq-rc